refactor(catalog): extract factory-droid model registry into a data module
Split the hand-maintained registry (types, table, base URLs, client version,
per-wire base URL map, model meta) out of the discovery logic into
factory-droid-models.ts. Drop seven dead registry fields, add per-model
reasoningReplay classification, fix kimi-k3 requiresEffort (explicit false
beats the canonical fill), and drop the test-only FACTORY_DROID_UPSTREAMS.
Rewrite the catalog test suite to contract-level assertions.
