+++ packages/catalog/src/provider-models/special.ts
@@ -3,7 +3,8 @@ import { buildModel } from "../build";
 import { apiRouteFor } from "../compat/behavior";
 import { type CodexModelDiscoveryResult, fetchCodexModels } from "../discovery/codex";
 import type { DevinModelDiscoveryOptions } from "../discovery/devin";
-import { buildFactoryDroidModel, FACTORY_DROID_MODELS, fetchFactoryDroidModels } from "../discovery/factory-droid";
+import { buildFactoryDroidModel, fetchFactoryDroidModels } from "../discovery/factory-droid";
+import { FACTORY_DROID_MODELS } from "../discovery/factory-droid-models";
 import { buildGitLabDuoWorkflowFallbackModel, fetchGitLabDuoWorkflowModels } from "../discovery/gitlab-duo-workflow";
 import type { ModelManagerOptions } from "../model-manager";
 import { getBundledModel } from "../models";
